Move UnitTest code path
