+++ app/utils/__tests__/process/makeWidgetWindow.spec.js
@@ -4,7 +4,7 @@ import { BrowserWindow as MockBrowserWindow } from 'app/__mocks__/electron';
 import storeMock from 'store/storeMain';
 import * as PATH from 'constants/path';
 import * as updateWidgetContentBounds from 'utils/process/updateWidgetContentBounds';
-import makeWidgetWindow from '../process/makeWidgetWindow';
+import makeWidgetWindow from '../../process/makeWidgetWindow';
 
 jest.mock('fs');
